Teryaqwithllm/requirements.txt
fastapi
uvicorn[standard]
psycopg2-binary
sqlalchemy
python-dotenv
Faker
pandas
hijri-converter
rapidfuzz
pyrebase4
